requirements.txt
Django==3.0.8
paho-mqtt==1.5.0
djangorestframework==3.11.0
#Scrapy==2.2.1
#scrapyd==1.2.1
#python-scrapyd-api==2.1.2
#itemadapter==0.1.0
